# Conflicts:
#	pet-adoption-system-web/src/views/user/stores/storesDetail.vue

+++ pet-adoption-system-api/src/main/java/cn/jasonone/mapper/PetInfoMapper.java
@@ -18,7 +18,7 @@ public interface PetInfoMapper {
 
     int insertSelective(PetInfo record);
 
-    PetInfo selectByPrimaryKey(Long id);
+    PetInfo selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(PetInfo record);
